Moved one test into another file
